Finalized version
Added comments fully explaining the code. Added error checks for the user's inputted webpage link, words to count, and choice of data visualization for the word counts. Decided to also have the program calculate the proportion of times that each word appeared in the text. This was then compared to the Bee Movie script's text. Program now extracts rows from the Wikipedia page's data frame with a larger and smaller proportion of words than the Bee Movie script's data frame to display to user. Updated visualizations to plot the Wikipedia page's and Bee Movie script's word counts on the same plot for users to make easier comparisons.